Keep the values as decimal

+++ src/ElectionKata/ElectionResult.cs
@@ -2,13 +2,13 @@
 {
     internal class ElectionResult
     {
-        internal ElectionResult(string party, double voteCount)
+        internal ElectionResult(string party, decimal voteCount)
         {
             Party = party;
             VoteCount = voteCount;
         }
 
         public string Party { get; }
-        public double VoteCount { get; }
+        public decimal VoteCount { get; }
     }
 }
